requirements.txt
streamlit>=1.44.0
supabase>=2.3.0
openai>=1.75.0
pandas>=2.2.0
matplotlib>=3.10.0
reportlab>=4.0.0
pillow>=11.0.0
python-dotenv>=1.0.0
httpx>=0.28.0
aiohttp>=3.11.0
pyjwt>=2.10.0
python-dateutil>=2.9.0
typing-extensions>=4.13.0
watchdog>=6.0.0
websockets>=14.0.0
requests>=2.32.0 